requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# API and Web Requests
requests>=2.31.0
httpx>=0.24.0

# Configuration
pyyaml>=6.0

# CLI
click>=8.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Data Processing
python-dateutil>=2.8.0

